fix an import problem issue of prepare_plot
